refactor: simplify preload-impl.ts
