--- packages/qwik/src/server/preload-strategy.ts
@@ -1,77 +1,0 @@
-import { getPlatform } from '@qwik.dev/core';
-import type { ResolvedManifest } from '@qwik.dev/core/optimizer';
-import { getSymbolHash } from './platform';
-import { flattenPrefetchResources } from './preload-utils';
-import { getQueue, preload, resetQueue } from './qwik-copy';
-import type { QRLInternal } from './qwik-types';
-import type { RenderToStringOptions } from './types';
-
-const getBundles = (qrls: QRLInternal[]) => {
-  const platform = getPlatform();
-  const bundles = (qrls as QRLInternal[])
-    ?.map((qrl) => {
-      const symbol = qrl.$symbol$;
-      const chunk = qrl.$chunk$;
-      const result = platform.chunkForSymbol(symbol, chunk, qrl.dev?.file);
-      if (result) {
-        return result[1];
-      }
-      return chunk;
-    })
-    .filter(Boolean) as string[];
-  return [...new Set(bundles)];
-};
-/** Returns paths to preload relative to the buildBase, with probabilities */
-export function getPreloadPaths(
-  qrls: QRLInternal[],
-  opts: RenderToStringOptions,
-  resolvedManifest: ResolvedManifest | undefined
-): string[] {
-  const prefetchStrategy = opts.prefetchStrategy;
-  if (prefetchStrategy === null) {
-    return [];
-  }
-  if (!resolvedManifest?.manifest.bundleGraph) {
-    return getBundles(qrls);
-  }
-
-  // TODO should we deprecate this?
-  if (typeof prefetchStrategy?.symbolsToPrefetch === 'function') {
-    // call user option symbolsToPrefetch()
-    try {
-      const prefetchResources = prefetchStrategy.symbolsToPrefetch({
-        manifest: resolvedManifest.manifest,
-      });
-      return flattenPrefetchResources(prefetchResources);
-    } catch (e) {
-      console.error('getPrefetchUrls, symbolsToPrefetch()', e);
-    }
-  }
-
-  // If we have a bundle graph, all we need is the symbols
-  const symbols = new Set<string>();
-  for (let i = 0; i < qrls.length; i++) {
-    const symbol = getSymbolHash(qrls[i].$symbol$);
-    if (symbol && symbol.length >= 10) {
-      symbols.add(symbol);
-    }
-  }
-  return [...symbols];
-}
-
-export const expandBundles = (names: string[], resolvedManifest?: ResolvedManifest) => {
-  if (!resolvedManifest?.manifest.bundleGraph) {
-    return [...new Set(names)];
-  }
-
-  resetQueue();
-
-  let probability = 0.99;
-  for (let i = 0; i < names.length; i++) {
-    preload(names[i], probability);
-    // later symbols have less probability
-    probability *= 0.95;
-  }
-
-  return getQueue();
-};
